Use 32-bit offsets in the string pool search tree to reduce its size

+++ pool.hpp
@@ -2,9 +2,9 @@
 #define POOL_HPP
 
 struct stringpool {
-	long long left;
-	long long right;
-	long long off;
+	unsigned long left;
+	unsigned long right;
+	unsigned long off;
 };
 
 long long addpool(struct memfile *poolfile, struct memfile *treefile, const char *s, char type);
